+++ cups/http-support.c
@@ -636,7 +636,11 @@ httpGetDateString(time_t t,		// I - Time in seconds
   struct tm	tdate;			// UNIX date/time data
 
 
-  gmtime_r(&t, &tdate);
+  if (!gmtime_r(&t, &tdate))
+  {
+    memset(&tdate, 0, sizeof(tdate));
+    tdate->tm_mday = 1;
+  }
 
   snprintf(s, slen, "%s, %02d %s %d %02d:%02d:%02d GMT", http_days[tdate.tm_wday], tdate.tm_mday, http_months[tdate.tm_mon], tdate.tm_year + 1900, tdate.tm_hour, tdate.tm_min, tdate.tm_sec);